@@ -28,50 +28,50 @@
     "@commitlint/cli": "^16.2.3",
     "@commitlint/config-conventional": "^16.2.1",
     "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.180",
-    "@types/node": "^17.0.21",
-    "@types/react": "^17.0.41",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.23",
+    "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
     "@types/styled-components": "^5.1.24",
-    "@typescript-eslint/eslint-plugin": "^5.15.0",
-    "@typescript-eslint/parser": "^5.15.0",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
     "@vitejs/plugin-react": "^1.2.0",
     "cross-env": "^7.0.3",
     "dotenv": "^16.0.0",
-    "eslint": "^8.11.0",
+    "eslint": "^8.12.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^16.1.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^26.1.2",
+    "eslint-plugin-jest": "^26.1.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react-hooks": "^4.4.0",
     "husky": "^7.0.4",
     "jest": "^27.5.1",
     "jest-svg-transformer": "^1.0.0",
     "lint-staged": "^12.3.7",
-    "prettier": "^2.6.0",
+    "prettier": "^2.6.1",
     "sass": "^1.49.9",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.6.2",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.3",
     "vite": "^2.8.6",
     "vite-plugin-compression": "^0.5.1",
-    "vite-plugin-svgr": "^1.0.1",
+    "vite-plugin-svgr": "^1.1.0",
     "vite-tsconfig-paths": "^3.4.1"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/user-event": "^14.0.0",
     "axios": "^0.26.1",
     "dayjs": "^1.11.0",
     "immer": "^9.0.12",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.2",
-    "styled-components": "^5.3.3",
+    "styled-components": "^5.3.5",
     "use-immer": "^0.6.0",
     "yup": "^0.32.11"
   },
